ExprType

Description

Expression Datatype

--------------------------------------

Wraps different operations in a

Expression Tree

Ops: Add - standard binary addition Mult - standard binary multiplication Const - wrapper for simple values Var - string identifier for vairables

getVars Function

--------------------------------------

Utilizes The Union Function

Helps Extract Variables As a List of Strings From a Given Expr Expression.

Documentation

data Expr a

Expression DataType Declaration

Constructors

Add (Expr a) (Expr a)
Standard Binary Addition

Mult (Expr a) (Expr a)
Standard Binary Multiplication

Const a
Value Wrapper

Var String
Variable Identifier

Cos (Expr a)
Cosine of Another Expression

Sin (Expr a)
Since of Another Expression

Natlog (Expr a)
Natural Log of Another Expression

Exp (Expr a)
Base e to The Power of Another Expression

Raise (Expr a) (Expr a)
Raise an Expression to The Power of Another Expression

getVars

Arguments

:: Expr a
Expr Type Expression as Input

-> [String]
List of Strings as Output

Function For Obtaining Variables From an Expr Type Expression